chore: refactor code layout, document plugin methods
