+++ package.json
@@ -22,7 +22,7 @@
   "author": "snyk.io",
   "license": "Apache-2.0",
   "dependencies": {
-    "@snyk/cli-interface": "^2.0.2",
+    "@snyk/cli-interface": "^2.0.3",
     "tmp": "0.0.33"
   },
   "devDependencies": {
