+++ pysrc/README.md
@@ -11,9 +11,10 @@ The entry point is `main` in `pip_resolve.py`.
 
 1. take pkg_resources.working_set (a list of all packages available in the current environment)
 2. convert it to a tree
-3. parse the manifest (requirements.txt/Pipfile) to find top-level deps
+3. parse the manifest (requirements.txt/Pipfile) to find the top-level deps
 4. select the parts of the tree that start from TLDs found in previous step
 5. determine actual installed versions for the packages in the tree
 6. convert that tree in DepTree format
 
 The parts 1 and 5 require access to the Python environment and thus have to be implemented in Python.
+The part 3, for requirements.txt, leverages the existing parsing library (pip).
