+++ test/build-args.test.js
@@ -1,21 +1,18 @@
 const test = require('tap').test;
-const plugin = require('../lib').__tests;
+const buildArgs = require('../lib/inspect-implementation').buildArgs;
 
 test('check build args with array', (t) => {
-  const result = plugin.buildArgs(
-    'requirements.txt',
-    false,
-    '../pysrc',
-    false,
-    ['-argOne', '-argTwo']
-  );
+  const result = buildArgs('requirements.txt', false, '../pysrc', false, [
+    '-argOne',
+    '-argTwo',
+  ]);
   t.match(result[0], /.*[/\\]pysrc[/\\]pip_resolve\.py/);
   t.deepEqual(result.slice(1), ['requirements.txt', '-argOne', '-argTwo']);
   t.end();
 });
 
 test('check build args with array & allowMissing', (t) => {
-  const result = plugin.buildArgs('requirements.txt', true, '../pysrc', false, [
+  const result = buildArgs('requirements.txt', true, '../pysrc', false, [
     '-argOne',
     '-argTwo',
   ]);
@@ -30,7 +27,7 @@ test('check build args with array & allowMissing', (t) => {
 });
 
 test('check build args with array & devDeps', (t) => {
-  const result = plugin.buildArgs('requirements.txt', false, '../pysrc', true, [
+  const result = buildArgs('requirements.txt', false, '../pysrc', true, [
     '-argOne',
     '-argTwo',
   ]);
@@ -45,7 +42,7 @@ test('check build args with array & devDeps', (t) => {
 });
 
 test('check build args with array & allowMissing & devDeps', (t) => {
-  const result = plugin.buildArgs('requirements.txt', true, '../pysrc', true, [
+  const result = buildArgs('requirements.txt', true, '../pysrc', true, [
     '-argOne',
     '-argTwo',
   ]);
@@ -61,7 +58,7 @@ test('check build args with array & allowMissing & devDeps', (t) => {
 });
 
 test('check build args with string', (t) => {
-  const result = plugin.buildArgs(
+  const result = buildArgs(
     'requirements.txt',
     false,
     '../pysrc',
@@ -74,7 +71,7 @@ test('check build args with string', (t) => {
 });
 
 test('check build args with string & allowMissing', (t) => {
-  const result = plugin.buildArgs(
+  const result = buildArgs(
     'requirements.txt',
     true,
     '../pysrc',
